refactor:전체적으로 컴포넌트 나눠서 폴더 분리
