+++ .prettierrc
@@ -2,9 +2,8 @@
   "semi": true,
   "singleQuote": false,
   "trailingComma": "all",
-  "printWidth": 80,
+  "printWidth": 70,
   "tabWidth": 2,
   "bracketSpacing": true,
-  "bracketSameLine": true,
   "arrowParens": "always"
 }
