+++ src/components/index.ts
@@ -1,0 +1,5 @@
+export { default as ComplexForm } from "./ComplexForm";
+export { default as Header } from "./Header";
+export { default as ItemList } from "./ItemList";
+export { default as Main } from "./Main";
+export { default as NotificationSystem } from "./NotificationSystem";
